backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
cryptography==41.0.5
python-dotenv==1.0.0